WIP
Added README + Data Controller.

Also created a RateMyChance Calculator...
